2008-05-09  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	Fixed infinite loop in checksum checking.
	* src/IteratableChecksumValidator.cc

+++ src/IteratableChecksumValidator.cc
@@ -84,7 +84,7 @@ void IteratableChecksumValidator::validateChunk()
 
 bool IteratableChecksumValidator::finished() const
 {
-  if((uint16_t)_currentOffset >= _dctx->getTotalLength()) {
+  if((uint64_t)_currentOffset >= _dctx->getTotalLength()) {
     _pieceStorage->getDiskAdaptor()->disableDirectIO();
     return true;
   } else {
